parsing fix WIP
