+++ Sources/Parsing/Quotes/manage_quotes.c
@@ -75,8 +75,8 @@ void	quote_found(t_dlist **temp, char *str, size_t *i, size_t *j)
 {
 	*j = *i + 1;
 	move_to_next_quote(str, i, str[(*i)]);
-	if (*j - *i)
-		split_into_dlst(temp, str, *i, *j);
+	// if (*j - *i)
+	split_into_dlst(temp, str, *i, *j);
 	if (str[(*i)])
 		(*i)++;
 	*j = *i;
@@ -105,32 +105,40 @@ char	*remove_quotes(char *str)
 	temp = NULL;
 	i = 0;
 	j = 0;
+	ft_dprintf(2, "bginning of rm quotes, str = %s\n", str);
 	while (str[i])
 	{
 		if (is_char_in_set(str[i], "\'\""))
-			quote_found(&temp, str, &i, &j);
+		{
+			j = i;
+			i++;
+			ft_dprintf(2, "in quotes before while i = %d, j = %d\n", i, j);
+			while (str[i] && str[i] != str[j])
+				i++;
+			ft_dprintf(2, "in quotes after while, i = %d, j = %d\n", i, j);
+			split_into_dlst(&temp, str, i, j + 1);
+			j = i;
+			i++;
+			
+		}
 		else
 		{
-			// if (not_a_quote_case(str, temp, &i, &j))
-			// {
-			// 	j = i;
-			// 	break ;
-			// }
+			// j = i;
+			// i++;
+			ft_dprintf(2, "in else before while, i = %d, j = %d\n", i, j);
 			while (str[i] && !is_char_in_set(str[i], "\'\""))
 				i++;
+			ft_dprintf(2, "in else after while, i = %d, j = %d\n", i, j);
 			split_into_dlst(&temp, str, i, j);
-			if (str[i])
-				i++;
-			else
-				break ;
+			// if (str[i])
+			// 	i++;
 			j = i;
-		}
+		}		
 	}
-	if (str[i])
-	{
-		i = ft_strlen(str);
-		split_into_dlst(&temp, str, i, j);
-	}
+	if (j != i)
+			split_into_dlst(&temp, str, i + 1, j - 1);
+	ft_dprintf(2, "end of rm quotes, lst =\n");
+	print_dlist(temp);
 	if (temp)
 	{
 		result = join_dlist(temp);
@@ -138,3 +146,51 @@ char	*remove_quotes(char *str)
 	}
 	return (NULL);
 }
+
+// char	*remove_quotes(char *str)
+// {
+// 	t_dlist	*temp;
+// 	char	*result;
+// 	size_t	i;
+// 	size_t	j;
+
+// 	temp = NULL;
+// 	i = 0;
+// 	j = 0;
+// 	while (str[i])
+// 	{
+// 		if (is_char_in_set(str[i], "\'\""))
+// 		{
+// 			quote_found(&temp, str, &i, &j);
+
+// 			ft_dprintf(2, "in in set quotes so far, lst =\n");
+// 			print_dlist(temp);
+// 		}
+// 		else
+// 		{
+// 			while (str[i] && !is_char_in_set(str[i], "\'\""))
+// 				i++;
+// 			split_into_dlst(&temp, str, i, j);
+// 			ft_dprintf(2, "in else so far, lst =\n");
+// 			print_dlist(temp);
+// 			if (str[i])
+// 				i++;
+// 			else
+// 				break ;
+// 			j = i;
+// 		}
+// 	}
+// 	if (str[i])
+// 	{
+// 		i = ft_strlen(str);
+// 		split_into_dlst(&temp, str, i, j);
+// 		ft_dprintf(2, "after while, so far, lst =\n");
+// 		print_dlist(temp);
+// 	}
+// 	if (temp)
+// 	{
+// 		result = join_dlist(temp);
+// 		return (free_dlist(&temp), result);
+// 	}
+// 	return (NULL);
+// }
