+++ Sources/Tokenizer/tokenizer.c
@@ -96,7 +96,7 @@ int	tokenizer(t_mshell *shell)
 	split_tokens_into_dlst(&shell->tok_lst, shell);
 	get_fds(shell, &shell->tok_lst);
 	create_cmd_arr(&shell->tok_lst, shell);
-	//manage_quotes_arr(&shell->tok_lst);
+	manage_quotes_arr(&shell->tok_lst);
 	free_tokens_dlist(&shell->tok_lst);
 	give_type(&shell->tok_lst);
 	if (!shell->tok_lst || !shell->tok_lst->cmd_arr
